# Backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hospital Token Management API",
    "version": "1.0.0",
    "description": "API Documentation for Hospital Token Management SaaS"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/patient/search": {
      "get": {
        "summary": "Search patient by phone",
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "phone",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Patient found"
          }
        }
      }
    },
    "/api/patient/{id}/history": {
      "get": {
        "summary": "Get patient medical history",
        "tags": [
          "Patient"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of previous visits"
          }
        }
      }
    },
...
  },
...
}